fix buggy assert in SSTableUtils.  patch by jbellis
git-svn-id: https://svn.apache.org/repos/asf/cassandra/trunk@955833 13f79535-47bb-0310-9956-ffa450edef68

+++ test/unit/org/apache/cassandra/io/sstable/SSTableUtils.java
@@ -52,9 +52,9 @@ public class SSTableUtils
         File tabledir = new File(tempdir, tablename);
         tabledir.mkdir();
         tabledir.deleteOnExit();
-        File datafile = new File(new Descriptor(tabledir, tablename, cfname, 0,
-                                                        false).filenameFor("Data.db"));
-        assert datafile.createNewFile();
+        File datafile = new File(new Descriptor(tabledir, tablename, cfname, 0, false).filenameFor("Data.db"));
+        if (!datafile.createNewFile())
+            throw new IOException("unable to create file " + datafile);
         datafile.deleteOnExit();
         return datafile;
     }
